Reorganize website demo files
