+++ website/src/pages/react-querybuilder/_constants/musicalInstruments.ts
