+++ website/src/pages/react-querybuilder/_styles/rqb-antd.less
