+++ website/src/pages/react-querybuilder/_styles/rqb-material.scss
@@ -10,7 +10,7 @@
     justify-content: center;
 
     & > div {
-      height: 33.33%;
+      height: 50%;
     }
   }
 }
