+++ website/src/pages/react-querybuilder/bulma.tsx
@@ -2,16 +2,18 @@ import BrowserOnly from '@docusaurus/BrowserOnly';
 import { QueryBuilderBulma } from '@react-querybuilder/bulma';
 import Layout from '@theme/Layout';
 import React from 'react';
-import './demo.scss';
-import './rqb-bulma.scss';
+import './_styles/demo.scss';
+import './_styles/rqb-bulma.scss';
 
 export default function ReactQueryBuilderDemo_Bulma() {
   return (
     <Layout description="React Query Builder Bulma Demo">
       <BrowserOnly fallback={<div>Loading...</div>}>
         {() => {
-          // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/consistent-type-imports
-          const Demo: typeof import('./_Demo').default = require('./_Demo').default;
+          // eslint-disable-next-line @typescript-eslint/consistent-type-imports
+          const Demo: typeof import('./_components/Demo').default =
+            // eslint-disable-next-line @typescript-eslint/no-var-requires
+            require('./_components/Demo').default;
           return (
             <QueryBuilderBulma>
               <Demo variant="bulma" />
